# remove strings.ToLower(s) ,avoid  redundant loops