there are loops  of   string `s` in function `strings.ToLower(s)` . And  function `normalizeScheme()` traverses the same string ,again. I think it  is  unnecessary.  removing strings.ToLower(s)  can  avoid  redundant loops.