@abhinav 2023-03-05
Thanks for the contribution, @tylitianrui.
Unfortunately, this change increases the number of allocations, rather than decreate.
I wrote a quick benchmark and ran it before and after this change:

```go
func BenchmarkNormalizeScheme(b *testing.B) {
	tests := []string{
		"file",
		"FILE",
		"FiLe",
	}

	b.ReportAllocs()
	b.ResetTimer()
	for i := 0; i < b.N; i++ {
		for _, tt := range tests {
			normalizeScheme(tt)
		}
	}
}
```

Here's the before/after:

```
name               old time/op    new time/op    delta
NormalizeScheme-2     241ns ± 7%     272ns ± 7%   ~     (p=0.100 n=3+3)

name               old alloc/op   new alloc/op   delta
NormalizeScheme-2     8.00B ± 0%    28.00B ± 0%   ~     (p=0.100 n=3+3)

name               old allocs/op  new allocs/op  delta
NormalizeScheme-2      2.00 ± 0%      4.00 ± 0%   ~     (p=0.100 n=3+3)
```

Allocations go up by two per operation.
```